xdiff: use usize instead of u32 in structs record, histindex, region

+++ xdiff/xhistogram.c
@@ -50,7 +50,7 @@
 #define LINE_END_PTR(n) (*line##n + *count##n - 1)
 
 struct record {
-	u32 ptr, cnt;
+	usize ptr, cnt;
 	struct record *next;
 };
 
@@ -62,16 +62,16 @@ struct histindex {
 	struct ivec_record_ptr record;
 	struct ivec_record_ptr line_map; /* map of line to record chain */
 	struct ivec_u32 next_ptrs;
-	u32 table_bits;
-	u32 max_chain_length;
-	u32 ptr_shift;
-	u32 cnt;
+	usize table_bits;
+	usize max_chain_length;
+	usize ptr_shift;
+	usize cnt;
 	bool has_common;
 };
 
 struct region {
-	u32 begin1, end1;
-	u32 begin2, end2;
+	usize begin1, end1;
+	usize begin2, end2;
 };
 
 #define LINE_MAP(i, a) (i->line_map.ptr[(a) - i->ptr_shift])
